+++ governance/ui/src/components/CouncilInformation/CouncilInformation.tsx
@@ -9,21 +9,23 @@ export default function CouncilInformation({ activeCouncil }: { activeCouncil: C
 
   return (
     <Flex py="5" maxW="735px" w="100%">
-      <CouncilImage
-        imageUrl={council?.image || ''}
-        width="130px"
-        height="130px"
-        imageProps={{ w: '108px', h: '108px' }}
-      />
-      <Flex flexDir="column" maxWidth="80%">
-        <Heading fontSize="lg" mb="2">
+      <Flex>
+        <CouncilImage
+          imageUrl={council?.image || ''}
+          width="130px"
+          height="130px"
+          imageProps={{ w: '108px', h: '108px' }}
+        />
+      </Flex>
+      <Flex flexDir="column" ml="24px">
+        <Heading fontSize="lg" mb="8px">
           {council?.title}
         </Heading>
-        <Text fontSize="sm" color="gray.500">
+        <Text w="581px" fontSize="14px" lineHeight="20px" fontFamily="heading" color="gray.500">
           {council?.description}
         </Text>
-        <Flex justifyContent="space-between" w="100%" mt="2">
-          <Heading fontSize="xs" mr="11">
+        <Flex w="100%" mt="8px">
+          <Heading fontSize="xs" mr="44px" lineHeight="110%">
             Council Seats: {getCouncilMembers?.length}
           </Heading>
           <Heading fontSize="xs">Stipends P/M: {council?.stipends}</Heading>
