+++ governance/ui/src/components/UserActionBox/UserActionBox.tsx
@@ -58,6 +58,7 @@ export default function UserActionBox({ activeCouncil }: UserActionBoxProps) {
         borderColor="gray.900"
         bg="navy.700"
         rounded="base"
+        mt="24px"
       >
         <Text
           w="225px"
